Add updated html file with project summary
After changes in the .ipynb file, I updated the html summary of the solution to include the changes. All code changes happened in step 5.
In particular, changes include:

1) Removal of redundant model training: previous code trained the model with backpropagation twice in step 5 (building the CNN from scratch)

2) Loaded the best set of coefficient on the right place: I cut-pasted the snippet of the optimal model coefficient loading on a separate code box for correctness. It is loading the coefficients before the model is compiled. 

The results are the same, after these changes. The errors fixed were mostly conceptual and weren't following best practice. 
See step 5 of the dog_app.ipynb for the corresponding code changes. 
See step 5 of the dog_app.html for these updates.